feat(widget): Rename 9dof widget to 9dof-widget
