+++ src/widgets/9dof-widget/index.ts
@@ -2,6 +2,6 @@ import { Widget } from '@wuespace/telestion-client-types';
 import { Widget as WidgetRenderer } from './widget';
 
 export const widget: Widget = {
-	name: '9dof',
+	name: '9dof-widget',
 	Widget: WidgetRenderer
 };
