+++ src/widgets/9dof-widget/model.ts
