+++ src/widgets/9dof-widget/table.tsx
