+++ src/widgets/index.ts
@@ -7,13 +7,13 @@ using the tc-cli generate widget command.
 
 import { Widget } from '@wuespace/telestion-client-types';
 import { widget as sampleWidget } from './sample-widget';
-import { widget as ninedof } from './9dof';
+import { widget as nineDofWidget } from './9dof-widget';
 import { widget as graphWidget } from './graph-widget';
 // IMPORT_INSERT_MARK
 
 export const projectWidgets: Widget[] = [
 	// ARRAY_FIRST_ELEMENT_INSERT_MARK
 	graphWidget as Widget,
-	ninedof,
+	nineDofWidget,
 	sampleWidget
 ];
